+++ julesTk/utils/font.py
@@ -1,0 +1,5 @@
+"""Provides class for working with font styles"""
+
+
+class Font(object):
+    """Combine size, family and more options in one class"""